LLPC code refine - Add ShaderModuleHelper for shader module related functions
1. Move shader module realted function to new  class ShaderModuleHelper
2. Move BinaryIStream and BinaryOStream to new file llpcBinaryStream
3. Remove unused function GetPipelineStatistics
4. Move MergeElfBinary to ElfWriter
5. Move TranslateSpirvToLlvm to llpcSpirvLowerTranslator
