+++ make/Makefile.llpc
@@ -12,6 +12,7 @@ LCXXINCS += -I$(LLPC_DEPTH)/patch/gfx9
 LCXXINCS += -I$(LLPC_DEPTH)/patch/gfx6/chip
 LCXXINCS += -I$(LLPC_DEPTH)/patch/gfx9/chip
 LCXXINCS += -I$(LLPC_DEPTH)/util
+LCXXINCS += -I$(LLPC_DEPTH)/translator/lib/SPIRV/libSPIRV
 
 ifeq ($(LLPC_BUILD_VEGA20), 1)
     LCXXDEFS += -DLLPC_BUILD_VEGA20=1 -DCHIP_HDR_VEGA20=1
@@ -49,7 +50,6 @@ ifeq ($(ICD_BUILD_LLPC), 1)
     LCXXINCS += -I$(LLPC_DEPTH)/patch/generate
     LCXXINCS += -I$(LLPC_DEPTH)/translator/include
     LCXXINCS += -I$(LLPC_DEPTH)/translator/lib/SPIRV
-    LCXXINCS += -I$(LLPC_DEPTH)/translator/lib/SPIRV/libSPIRV
     LCXXINCS += -I$(LLPC_DEPTH)/translator/lib/SPIRV/Mangler
 
     vpath %.cpp $(LLPC_DEPTH)/builder
@@ -186,6 +186,7 @@ endif
         llpcPassManager.cpp                 \
         llpcPipelineDumper.cpp              \
         llpcPipelineShaders.cpp             \
+        llpcShaderModuleHelper.cpp          \
         llpcStartStopTimer.cpp              \
         llpcTimerProfiler.cpp               \
         llpcUtil.cpp
