+++ util/llpcInternal.cpp
@@ -304,60 +304,6 @@ void SetShaderStageToModule(
 }
 
 // =====================================================================================================================
-// Converts the SPIR-V execution model to the shader stage
-ShaderStage ConvertToStageShage(
-    uint32_t execModel)  // SPIR-V execution model
-{
-    switch (execModel)
-    {
-    case spv::ExecutionModelVertex:
-    case spv::ExecutionModelTessellationControl:
-    case spv::ExecutionModelTessellationEvaluation:
-    case spv::ExecutionModelGeometry:
-    case spv::ExecutionModelFragment:
-    case spv::ExecutionModelGLCompute:
-        {
-            return static_cast<ShaderStage>(execModel);
-        }
-    case spv::ExecutionModelCopyShader:
-        {
-            return ShaderStageCopyShader;
-        }
-    }
-
-    LLPC_NEVER_CALLED();
-    return ShaderStageInvalid;
-}
-
-// =====================================================================================================================
-// Converts the shader stage to the SPIR-V execution model
-spv::ExecutionModel ConvertToExecModel(
-    ShaderStage shaderStage)  // Shader stage
-{
-    switch (shaderStage)
-    {
-    case ShaderStageVertex:
-    case ShaderStageTessControl:
-    case ShaderStageTessEval:
-    case ShaderStageGeometry:
-    case ShaderStageFragment:
-    case ShaderStageCompute:
-        {
-            return static_cast<spv::ExecutionModel>(shaderStage);
-        }
-    case ShaderStageCopyShader:
-        {
-            return spv::ExecutionModelCopyShader;
-        }
-    default:
-        {
-            LLPC_NEVER_CALLED();
-            return static_cast <spv::ExecutionModel>(0);
-        }
-    }
-}
-
-// =====================================================================================================================
 // Gets the shader stage from the specified calling convention.
 ShaderStage GetShaderStageFromCallingConv(
     uint32_t        stageMask,  // Shader stage mask for the pipeline
@@ -442,206 +388,6 @@ bool CanBitCast(
     }
 
     return valid;
-}
-
-// =====================================================================================================================
-// Checks whether input binary data is SPIR-V binary
-bool IsSpirvBinary(
-    const BinaryData*  pShaderBin)  // [in] Shader binary codes
-{
-    bool isSpvBinary = false;
-    if (pShaderBin->codeSize > sizeof(SpirvHeader))
-    {
-        const SpirvHeader* pHeader = reinterpret_cast<const SpirvHeader*>(pShaderBin->pCode);
-        if ((pHeader->magicNumber == MagicNumber) && (pHeader->spvVersion <= spv::Version) && (pHeader->reserved == 0))
-        {
-            isSpvBinary = true;
-        }
-    }
-
-    return isSpvBinary;
-}
-
-// =====================================================================================================================
-// Checks whether input binary data is LLVM bitcode.
-bool IsLlvmBitcode(
-    const BinaryData*  pShaderBin)  // [in] Shader binary codes
-{
-    static uint32_t BitcodeMagicNumber = 0xDEC04342; // 0x42, 0x43, 0xC0, 0xDE
-    bool isLlvmBitcode = false;
-    if ((pShaderBin->codeSize > 4) &&
-        (*reinterpret_cast<const uint32_t*>(pShaderBin->pCode) == BitcodeMagicNumber))
-    {
-        isLlvmBitcode = true;
-    }
-
-    return isLlvmBitcode;
-}
-
-// =====================================================================================================================
-// Gets the shader stage mask from the SPIR-V binary according to the specified entry-point.
-//
-// Returns 0 on error, or the stage mask of the specified entry-point on success.
-uint32_t GetStageMaskFromSpirvBinary(
-    const BinaryData*  pSpvBin,      // [in] SPIR-V binary
-    const char*        pEntryName)   // [in] Name of entry-point
-{
-    uint32_t stageMask = 0;
-
-    const uint32_t* pCode = reinterpret_cast<const uint32_t*>(pSpvBin->pCode);
-    const uint32_t* pEnd  = pCode + pSpvBin->codeSize / sizeof(uint32_t);
-
-    if (IsSpirvBinary(pSpvBin))
-    {
-        // Skip SPIR-V header
-        const uint32_t* pCodePos = pCode + sizeof(SpirvHeader) / sizeof(uint32_t);
-
-        while (pCodePos < pEnd)
-        {
-            uint32_t opCode = (pCodePos[0] & OpCodeMask);
-            uint32_t wordCount = (pCodePos[0] >> WordCountShift);
-
-            if ((wordCount == 0) || (pCodePos + wordCount > pEnd))
-            {
-                LLPC_ERRS("Invalid SPIR-V binary\n");
-                stageMask = 0;
-                break;
-            }
-
-            if (opCode == OpEntryPoint)
-            {
-                LLPC_ASSERT(wordCount >= 4);
-
-                // The fourth word is start of the name string of the entry-point
-                const char* pName = reinterpret_cast<const char*>(&pCodePos[3]);
-                if (strcmp(pEntryName, pName) == 0)
-                {
-                    // An matching entry-point is found
-                    stageMask |= ShaderStageToMask(ConvertToStageShage(pCodePos[1]));
-                }
-            }
-
-            // All "OpEntryPoint" are before "OpFunction"
-            if (opCode == OpFunction)
-            {
-                break;
-            }
-
-            pCodePos += wordCount;
-        }
-    }
-    else
-    {
-        LLPC_ERRS("Invalid SPIR-V binary\n");
-    }
-
-    return stageMask;
-}
-
-// =====================================================================================================================
-// Gets the entry-point name from the SPIR-V binary
-//
-// NOTE: This function is for single entry-point. If the SPIR-V binary contains multiple entry-points, we get the name
-// of the first entry-point and ignore others.
-const char* GetEntryPointNameFromSpirvBinary(
-    const BinaryData* pSpvBin) // [in] SPIR-V binary
-{
-    const char* pEntryName = nullptr;
-
-    const uint32_t* pCode = reinterpret_cast<const uint32_t*>(pSpvBin->pCode);
-    const uint32_t* pEnd = pCode + pSpvBin->codeSize / sizeof(uint32_t);
-
-    if (IsSpirvBinary(pSpvBin))
-    {
-        // Skip SPIR-V header
-        const uint32_t* pCodePos = pCode + sizeof(SpirvHeader) / sizeof(uint32_t);
-
-        while (pCodePos < pEnd)
-        {
-            uint32_t opCode = (pCodePos[0] & OpCodeMask);
-            uint32_t wordCount = (pCodePos[0] >> WordCountShift);
-
-            if ((wordCount == 0) || (pCodePos + wordCount > pEnd))
-            {
-                LLPC_ERRS("Invalid SPIR-V binary\n");
-                break;
-            }
-
-            if (opCode == OpEntryPoint)
-            {
-                LLPC_ASSERT(wordCount >= 4);
-
-                // The fourth word is start of the name string of the entry-point
-                pEntryName = reinterpret_cast<const char*>(&pCodePos[3]);
-                break;
-            }
-
-            // All "OpEntryPoint" are before "OpFunction"
-            if (opCode == OpFunction)
-            {
-                break;
-            }
-
-            pCodePos += wordCount;
-        }
-
-        if (pEntryName == nullptr)
-        {
-            LLPC_ERRS("Entry-point not found\n");
-            pEntryName = "";
-        }
-    }
-    else
-    {
-        LLPC_ERRS("Invalid SPIR-V binary\n");
-        pEntryName = "";
-    }
-
-    return pEntryName;
-}
-
-// =====================================================================================================================
-// Verifies if the SPIR-V binary is valid and is supported
-Result VerifySpirvBinary(
-    const BinaryData* pSpvBin)  // [in] SPIR-V binary
-{
-    Result result = Result::Success;
-
-#define _SPIRV_OP(x,...) Op##x,
-    static const std::set<Op> OpSet{
-       {
-        #include "SPIRVOpCodeEnum.h"
-       }
-    };
-#undef _SPIRV_OP
-
-    const uint32_t* pCode = reinterpret_cast<const uint32_t*>(pSpvBin->pCode);
-    const uint32_t* pEnd  = pCode + pSpvBin->codeSize / sizeof(uint32_t);
-
-    // Skip SPIR-V header
-    const uint32_t* pCodePos = pCode + sizeof(SpirvHeader) / sizeof(uint32_t);
-
-    while (pCodePos < pEnd)
-    {
-        Op opCode = static_cast<Op>(pCodePos[0] & OpCodeMask);
-        uint32_t wordCount = (pCodePos[0] >> WordCountShift);
-
-        if ((wordCount == 0) || (pCodePos + wordCount > pEnd))
-        {
-            result = Result::ErrorInvalidShader;
-            break;
-        }
-
-        if (OpSet.find(opCode) == OpSet.end())
-        {
-            result = Result::ErrorInvalidShader;
-            break;
-        }
-
-        pCodePos += wordCount;
-    }
-
-    return result;
 }
 
 // =====================================================================================================================
