+++ util/llpcUtil.cpp
@@ -35,6 +35,9 @@
 #include "llpcDebug.h"
 #include "llpcElfReader.h"
 #include "llpcUtil.h"
+
+#include "spirvExt.h"
+
 #include "palPipelineAbi.h"
 
 #if defined(_WIN32)
@@ -138,6 +141,60 @@ uint32_t ShaderStageToMask(
 {
     LLPC_ASSERT((stage < ShaderStageCount) || (stage == ShaderStageCopyShader));
     return (1 << stage);
+}
+
+// =====================================================================================================================
+// Converts the SPIR-V execution model to the shader stage
+ShaderStage ConvertToStageShage(
+    uint32_t execModel)  // SPIR-V execution model
+{
+    switch (execModel)
+    {
+    case spv::ExecutionModelVertex:
+    case spv::ExecutionModelTessellationControl:
+    case spv::ExecutionModelTessellationEvaluation:
+    case spv::ExecutionModelGeometry:
+    case spv::ExecutionModelFragment:
+    case spv::ExecutionModelGLCompute:
+        {
+            return static_cast<ShaderStage>(execModel);
+        }
+    case spv::ExecutionModelCopyShader:
+        {
+            return ShaderStageCopyShader;
+        }
+    }
+
+    LLPC_NEVER_CALLED();
+    return ShaderStageInvalid;
+}
+
+// =====================================================================================================================
+// Converts the shader stage to the SPIR-V execution model
+spv::ExecutionModel ConvertToExecModel(
+    ShaderStage shaderStage)  // Shader stage
+{
+    switch (shaderStage)
+    {
+    case ShaderStageVertex:
+    case ShaderStageTessControl:
+    case ShaderStageTessEval:
+    case ShaderStageGeometry:
+    case ShaderStageFragment:
+    case ShaderStageCompute:
+        {
+            return static_cast<spv::ExecutionModel>(shaderStage);
+        }
+    case ShaderStageCopyShader:
+        {
+            return spv::ExecutionModelCopyShader;
+        }
+    default:
+        {
+            LLPC_NEVER_CALLED();
+            return static_cast <spv::ExecutionModel>(0);
+        }
+    }
 }
 
 // =====================================================================================================================
